EDMT-215 refactor : mock 파일 구조 분리 및 잘못된 경로 수정
