+++ src/mocks/handlers/notice/get-notice-detail.ts
