+++ src/mocks/handlers/student-manage/create-record-detail.ts
