+++ src/mocks/handlers/student-manage/delete-record-detail.ts
