+++ src/mocks/handlers/student-manage/patch-record-detail.ts
