+++ src/mocks/handlers/write-record/get-summary-record-detail.ts
