+++ src/mocks/handlers/write-record/post-summary-record-detail.ts
