+++ src/services/auth/get-verify-email.ts
@@ -1,7 +1,6 @@
 import type { Response } from '@/types/api/response';
 
 export const getVerifyEmail = async (id: string, code: string): Promise<Response<null>> => {
-  console.log(id, code);
   const res = await fetch(
     `${process.env.API_URL}/api/v1/auth/verify-email?id=${encodeURIComponent(id)}&code=${encodeURIComponent(code)}`,
     {
